Differentiate Gradle and indexing target inside the generator task
