--- subprojects/ide-native/src/main/java/org/gradle/ide/xcode/internal/GradleBuildTarget.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.ide.xcode.internal;
-
-public class GradleBuildTarget extends XcodeTarget {
-    private String taskName;
-    private String gradleCommand;
-
-    public GradleBuildTarget(String name) {
-        super(name);
-    }
-
-    public String getTaskName() {
-        return taskName;
-    }
-
-    public void setTaskName(String taskName) {
-        this.taskName = taskName;
-    }
-
-    public String getGradleCommand() {
-        return gradleCommand;
-    }
-
-    public void setGradleCommand(String gradleCommand) {
-        this.gradleCommand = gradleCommand;
-    }
-}
