--- subprojects/ide-native/src/main/java/org/gradle/ide/xcode/internal/IndexingSwiftTarget.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.ide.xcode.internal;
-
-import java.io.File;
-import java.util.HashSet;
-import java.util.Set;
-
-public class IndexingSwiftTarget extends XcodeTarget {
-    private final Set<File> sources = new HashSet<File>();
-
-    public IndexingSwiftTarget(String name) {
-        super(name);
-    }
-
-    public Set<File> getSources() {
-        return sources;
-    }
-
-    public void setSources(Set<File> sources) {
-        this.sources.addAll(sources);
-    }
-}
